+++ CHANGES.rst
@@ -3,8 +3,44 @@ Changelog
 
 A list of changes between each release
 
-0.9.0.dev (Development Version)
+0.10.0 (2018-10-16)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+- Moved all API calls to own module for easier maintainability
+- Added network ids to sync module and cameras to allow for multi-network use
+- Removed dependency on video existance prior to camera setup (fixes `#93 <https://github.com/fronzbot/blinkpy/issues/#93>`_)
+- Camera wifi_strength now reported in wifi "bars" rather than dBm due to API endpoint change
+- Use homescreen thumbnail as fallback in case it's not in the camera endpoint
+- Removed "armed" and "status" attributes from camera (status of camera only reported by "motion_enabled" now)
+- Added serial number attributes to sync module and cameras
+- Check network_id from login response and verify that network is onboarded (fixes `#90 <https://github.com/fronzbot/#90>`_)
+- Check if retrieved clip is "None" prior to storing in cache
+
+0.9.0 (2018-09-27)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+- Complete code refactoring to enable future multi-sync module support
+- Add image and video caching to the cameras
+- Add internal throttling of system refresh
+- Use session for http requests
+
+**Breaking change:**
+- Cameras now accessed through sync module Blink.sync.cameras
+
+
+0.8.1 (2018-09-24)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+- Update requirements_test.txt
+- Update linter versions
+- Fix pylint warnings
+  - Remove object from class declarations
+  - Remove useless returns from functions
+- Fix pylint errors
+  - change if comparison to fix (consider-using-in)
+  - Disabled no else-if-return check
+- Fix useless-import-alias
+- Disable no-else-return
+- Fix motion detection
+  - Use an array of recent video clips to determine if motion has been detected.
+  - Reset the value every system refresh
 
 0.8.0 (2018-05-21)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^